+++ standalone/include/psen_scan_v2_standalone/data_conversion_layer/laserscan_conversions.h
@@ -16,9 +16,10 @@
 #ifndef PSEN_SCAN_V2_STANDALONE_LASERSCAN_CONVERSIONS_H
 #define PSEN_SCAN_V2_STANDALONE_LASERSCAN_CONVERSIONS_H
 
+#include "psen_scan_v2_standalone/configuration/default_parameters.h"
 #include "psen_scan_v2_standalone/data_conversion_layer/angle_conversions.h"
-#include "psen_scan_v2_standalone/laserscan.h"
 #include "psen_scan_v2_standalone/data_conversion_layer/monitoring_frame_msg.h"
+#include "psen_scan_v2_standalone/laserscan.h"
 
 #include <algorithm>
 #include <numeric>
@@ -52,46 +53,62 @@ public:
    * @see data_conversion_layer::monitoring_frame::Message
    * @see ScannerV2
    */
-  static LaserScan toLaserScan(const std::vector<data_conversion_layer::monitoring_frame::Message>& frames);
+  static LaserScan
+  toLaserScan(const std::vector<data_conversion_layer::monitoring_frame::MessageStamped>& stamped_msgs);
 
 private:
-  static std::vector<int>
-  getFilledFramesIndicesSortedByThetaAngle(const std::vector<data_conversion_layer::monitoring_frame::Message>& frames);
+  static std::vector<int> getFilledFramesIndicesSortedByThetaAngle(
+      const std::vector<data_conversion_layer::monitoring_frame::MessageStamped>& stamped_msgs);
   static util::TenthOfDegree
-  calculateMaxAngle(const std::vector<data_conversion_layer::monitoring_frame::Message>& frames,
+  calculateMaxAngle(const std::vector<data_conversion_layer::monitoring_frame::MessageStamped>& stamped_msgs,
                     const util::TenthOfDegree& min_angle);
-  static void validateMonitoringFrames(const std::vector<data_conversion_layer::monitoring_frame::Message>& frames,
-                                       const std::vector<int>& sorted_frames_indices);
-  static bool allResolutionsMatch(const std::vector<data_conversion_layer::monitoring_frame::Message>& frames);
-  static bool allScanCountersMatch(const std::vector<data_conversion_layer::monitoring_frame::Message>& frames);
-  static bool thetaAnglesFitTogether(const std::vector<data_conversion_layer::monitoring_frame::Message>& frames,
-                                     const std::vector<int>& sorted_frames_indices);
+  static int64_t
+  calculateTimestamp(const std::vector<data_conversion_layer::monitoring_frame::MessageStamped>& stamped_msgs,
+                     const std::vector<int>& filled_stamped_msgs_indices);
+  static int64_t calculateFirstRayTime(const data_conversion_layer::monitoring_frame::MessageStamped& stamped_msg);
+  static void
+  validateMonitoringFrames(const std::vector<data_conversion_layer::monitoring_frame::MessageStamped>& stamped_msgs,
+                           const std::vector<int>& sorted_stamped_msgs_indices);
+  static bool
+  allResolutionsMatch(const std::vector<data_conversion_layer::monitoring_frame::MessageStamped>& stamped_msgs);
+  static bool
+  allScanCountersMatch(const std::vector<data_conversion_layer::monitoring_frame::MessageStamped>& stamped_msgs);
+  static bool
+  thetaAnglesFitTogether(const std::vector<data_conversion_layer::monitoring_frame::MessageStamped>& stamped_msgs,
+                         const std::vector<int>& sorted_stamped_msgs_indices);
 };
 
-inline LaserScan
-LaserScanConverter::toLaserScan(const std::vector<data_conversion_layer::monitoring_frame::Message>& frames)
+inline LaserScan LaserScanConverter::toLaserScan(
+    const std::vector<data_conversion_layer::monitoring_frame::MessageStamped>& stamped_msgs)
 {
-  if (frames.empty())
+  if (stamped_msgs.empty())
   {
     throw ScannerProtocolViolationError("At least one monitoring frame is necessary to create a LaserScan");
   }
 
-  std::vector<int> sorted_frames_indices = getFilledFramesIndicesSortedByThetaAngle(frames);
-  validateMonitoringFrames(frames, sorted_frames_indices);
+  std::vector<int> sorted_stamped_msgs_indices = getFilledFramesIndicesSortedByThetaAngle(stamped_msgs);
+  validateMonitoringFrames(stamped_msgs, sorted_stamped_msgs_indices);
 
-  const auto min_angle = frames[sorted_frames_indices[0]].fromTheta();
-  const auto max_angle = calculateMaxAngle(frames, min_angle);
+  const auto min_angle = stamped_msgs[sorted_stamped_msgs_indices[0]].msg_.fromTheta();
+  const auto max_angle = calculateMaxAngle(stamped_msgs, min_angle);
+
+  const auto timestamp = calculateTimestamp(stamped_msgs, sorted_stamped_msgs_indices);
 
   std::vector<double> measurements;
   std::vector<double> intensities;
 
-  for (auto index : sorted_frames_indices)
+  for (auto index : sorted_stamped_msgs_indices)
   {
-    measurements.insert(measurements.end(), frames[index].measurements().begin(), frames[index].measurements().end());
-    intensities.insert(intensities.end(), frames[index].intensities().begin(), frames[index].intensities().end());
+    measurements.insert(measurements.end(),
+                        stamped_msgs[index].msg_.measurements().begin(),
+                        stamped_msgs[index].msg_.measurements().end());
+    intensities.insert(intensities.end(),
+                       stamped_msgs[index].msg_.intensities().begin(),
+                       stamped_msgs[index].msg_.intensities().end());
   }
 
-  LaserScan scan(frames[0].resolution(), min_angle, max_angle);
+  LaserScan scan(
+      stamped_msgs[0].msg_.resolution(), min_angle, max_angle, stamped_msgs[0].msg_.scanCounter(), timestamp);
   scan.setMeasurements(measurements);
   scan.setIntensities(intensities);
 
@@ -99,84 +116,109 @@ LaserScanConverter::toLaserScan(const std::vector<data_conversion_layer::monitor
 }
 
 inline std::vector<int> LaserScanConverter::getFilledFramesIndicesSortedByThetaAngle(
-    const std::vector<data_conversion_layer::monitoring_frame::Message>& frames)
+    const std::vector<data_conversion_layer::monitoring_frame::MessageStamped>& stamped_msgs)
 {
-  std::vector<int> sorted_filled_frames_indices(frames.size());
-  std::iota(sorted_filled_frames_indices.begin(), sorted_filled_frames_indices.end(), 0);
-  std::sort(sorted_filled_frames_indices.begin(), sorted_filled_frames_indices.end(), [frames](int i1, int i2) {
-    return frames[i1].fromTheta() < frames[i2].fromTheta();
-  });
+  std::vector<int> sorted_filled_stamped_msgs_indices(stamped_msgs.size());
+  std::iota(sorted_filled_stamped_msgs_indices.begin(), sorted_filled_stamped_msgs_indices.end(), 0);
+  std::sort(
+      sorted_filled_stamped_msgs_indices.begin(),
+      sorted_filled_stamped_msgs_indices.end(),
+      [stamped_msgs](int i1, int i2) { return stamped_msgs[i1].msg_.fromTheta() < stamped_msgs[i2].msg_.fromTheta(); });
 
   // The following contains a missing line in the coverage report, which does not make sense.
   // LCOV_EXCL_START
-  sorted_filled_frames_indices.erase(std::remove_if(sorted_filled_frames_indices.begin(),
-                                                    sorted_filled_frames_indices.end(),
-                                                    [frames](int i) { return frames[i].measurements().empty(); }),
-                                     sorted_filled_frames_indices.end());
+  sorted_filled_stamped_msgs_indices.erase(
+      std::remove_if(sorted_filled_stamped_msgs_indices.begin(),
+                     sorted_filled_stamped_msgs_indices.end(),
+                     [stamped_msgs](int i) { return stamped_msgs[i].msg_.measurements().empty(); }),
+      sorted_filled_stamped_msgs_indices.end());
   // LCOV_EXCL_STOP
 
-  return sorted_filled_frames_indices;
+  return sorted_filled_stamped_msgs_indices;
 }
 
-inline util::TenthOfDegree
-LaserScanConverter::calculateMaxAngle(const std::vector<data_conversion_layer::monitoring_frame::Message>& frames,
-                                      const util::TenthOfDegree& min_angle)
+inline util::TenthOfDegree LaserScanConverter::calculateMaxAngle(
+    const std::vector<data_conversion_layer::monitoring_frame::MessageStamped>& stamped_msgs,
+    const util::TenthOfDegree& min_angle)
 {
-  const auto resolution = frames[0].resolution();
-  const uint16_t number_of_samples =
-      std::accumulate(frames.begin(), frames.end(), uint16_t{ 0 }, [](uint16_t total, const auto& frame) {
-        return total + frame.measurements().size();
+  const auto resolution = stamped_msgs[0].msg_.resolution();
+  const uint16_t number_of_samples = std::accumulate(
+      stamped_msgs.begin(), stamped_msgs.end(), uint16_t{ 0 }, [](uint16_t total, const auto& stamped_msg) {
+        return total + stamped_msg.msg_.measurements().size();
       });
   return min_angle + resolution * static_cast<int>(number_of_samples - 1);
 }
 
-inline void LaserScanConverter::validateMonitoringFrames(
-    const std::vector<data_conversion_layer::monitoring_frame::Message>& frames,
-    const std::vector<int>& sorted_frames_indices)
+inline int64_t LaserScanConverter::calculateTimestamp(
+    const std::vector<data_conversion_layer::monitoring_frame::MessageStamped>& stamped_msgs,
+    const std::vector<int>& filled_stamped_msgs_indices)
 {
-  if (!allResolutionsMatch(frames))
+  const auto it = std::min_element(
+      filled_stamped_msgs_indices.begin(), filled_stamped_msgs_indices.end(), [&stamped_msgs](int i, int j) {
+        return stamped_msgs[i].stamp_ < stamped_msgs[j].stamp_;
+      });  // determines stamped_msg with smallest stamp
+  return calculateFirstRayTime(stamped_msgs[*it]);
+}
+
+inline int64_t
+LaserScanConverter::calculateFirstRayTime(const data_conversion_layer::monitoring_frame::MessageStamped& stamped_msg)
+{
+  const double time_per_scan_in_ns{ configuration::TIME_PER_SCAN_IN_S * 1000000000.0 };
+  const double scan_interval_in_degree{ stamped_msg.msg_.resolution().value() *
+                                        (stamped_msg.msg_.measurements().size() - 1) / 10.0 };
+  return stamped_msg.stamp_ - static_cast<int64_t>(std::round(scan_interval_in_degree * time_per_scan_in_ns / 360.0));
+}
+
+inline void LaserScanConverter::validateMonitoringFrames(
+    const std::vector<data_conversion_layer::monitoring_frame::MessageStamped>& stamped_msgs,
+    const std::vector<int>& sorted_stamped_msgs_indices)
+{
+  if (!allResolutionsMatch(stamped_msgs))
   {
     throw ScannerProtocolViolationError("The resolution of all monitoring frames has to be the same.");
   }
-  else if (!allScanCountersMatch(frames))
+  else if (!allScanCountersMatch(stamped_msgs))
   {
     throw ScannerProtocolViolationError("The scan counters of all monitoring frames have to be the same.");
   }
-  else if (!thetaAnglesFitTogether(frames, sorted_frames_indices))
+  else if (!thetaAnglesFitTogether(stamped_msgs, sorted_stamped_msgs_indices))
   {
     throw ScannerProtocolViolationError("The monitoring frame ranges do not cover the whole scan range");
   }
 }
 
-inline bool
-LaserScanConverter::allResolutionsMatch(const std::vector<data_conversion_layer::monitoring_frame::Message>& frames)
+inline bool LaserScanConverter::allResolutionsMatch(
+    const std::vector<data_conversion_layer::monitoring_frame::MessageStamped>& stamped_msgs)
 {
-  const auto resolution = frames[0].resolution();
-  return std::all_of(
-      frames.begin(), frames.end(), [resolution](const auto& frame) { return frame.resolution() == resolution; });
+  const auto resolution = stamped_msgs[0].msg_.resolution();
+  return std::all_of(stamped_msgs.begin(), stamped_msgs.end(), [resolution](const auto& stamped_msg) {
+    return stamped_msg.msg_.resolution() == resolution;
+  });
 }
 
-inline bool
-LaserScanConverter::allScanCountersMatch(const std::vector<data_conversion_layer::monitoring_frame::Message>& frames)
+inline bool LaserScanConverter::allScanCountersMatch(
+    const std::vector<data_conversion_layer::monitoring_frame::MessageStamped>& stamped_msgs)
 {
-  const auto scan_counter = frames[0].scanCounter();
-  return std::all_of(
-      frames.begin(), frames.end(), [scan_counter](const auto& frame) { return frame.scanCounter() == scan_counter; });
+  const auto scan_counter = stamped_msgs[0].msg_.scanCounter();
+  return std::all_of(stamped_msgs.begin(), stamped_msgs.end(), [scan_counter](const auto& stamped_msg) {
+    return stamped_msg.msg_.scanCounter() == scan_counter;
+  });
 }
 
-inline bool
-LaserScanConverter::thetaAnglesFitTogether(const std::vector<data_conversion_layer::monitoring_frame::Message>& frames,
-                                           const std::vector<int>& sorted_filled_frames_indices)
+inline bool LaserScanConverter::thetaAnglesFitTogether(
+    const std::vector<data_conversion_layer::monitoring_frame::MessageStamped>& stamped_msgs,
+    const std::vector<int>& sorted_filled_stamped_msgs_indices)
 {
-  util::TenthOfDegree lastEnd = frames[sorted_filled_frames_indices[0]].fromTheta();
-  for (auto index : sorted_filled_frames_indices)
+  util::TenthOfDegree lastEnd = stamped_msgs[sorted_filled_stamped_msgs_indices[0]].msg_.fromTheta();
+  for (auto index : sorted_filled_stamped_msgs_indices)
   {
-    const auto& frame = frames[index];
-    if (lastEnd != frame.fromTheta())
+    const auto& stamped_msg = stamped_msgs[index];
+    if (lastEnd != stamped_msg.msg_.fromTheta())
     {
       return false;
     }
-    lastEnd = frame.fromTheta() + frame.resolution() * static_cast<int>(frame.measurements().size());
+    lastEnd = stamped_msg.msg_.fromTheta() +
+              stamped_msg.msg_.resolution() * static_cast<int>(stamped_msg.msg_.measurements().size());
   }
   return true;
 }