+++ test/include/psen_scan_v2/laserscan_validator.h
@@ -17,6 +17,7 @@
 
 #include <algorithm>
 #include <atomic>
+#include <chrono>
 #include <future>
 #include <map>
 #include <stdexcept>